+++ changes.md
@@ -1,4 +1,4 @@
-## V0.1.2
+## V0.1.1
 * Added better logging for combat events to hopefully track down a duplicate announcement bug.
 * Fixed incorrect localization lookups for map nodes and merchant slots.
 * Map nodes now only announce traveled state (IE you have been there before); the reachable and unreachable state announcements were irrelevant and causing confusion.
@@ -7,4 +7,6 @@
 * Added events and announcements for card upgrades and when cards are added to your deck.
 * Mod settings menus are now sorted alphabetically.
 * The default keyboard binding for back is now backspace instead of escape to avoid input conflicts. You will have to adjust this keybinding yourself if you have already played the game though as your settings file doesn't reset to defaults on mod update.
-* Fixed an issue where the focus would glitch during the player's end of turn sequence, resulting in readouts of random controls.
+* Fixed an issue where the focus would glitch during the player's end of turn sequence resulting in readouts of random controls.
+* Reordered card label to read as "{name}, {cost}, {type}"; also added setting for verbose costs (when unchecked costs simply read as numbers with no label.)
+* Added generic mod settings attribute system for registering configurable settings on any class (events, UI elements, etc.)